containers/chatbot/requirements.txt
boto3==1.26.84
openai==0.27.0
discord.py==2.2.2
